requirements.txt
Flask==2.0.1
requests==2.25.1
gunicorn==20.1.0
Flask-SocketIO==5.2.0
python-dotenv==0.19.2
python-dateutil==2.8.2
reportlab==3.5.67
